+++ wbsModels/BudBurst/BudBurstSaintAmant.cpp
@@ -571,7 +571,7 @@ namespace WBSF
 	{
 		CTPeriod p(result.m_ref.GetYear(), JANUARY, DAY_01, result.m_ref.GetYear(), DECEMBER, DAY_31);
 		int pos = output.GetFirstIndex(O_SDI, ">", result.m_obs[0], 0, p);
-		return pos >= 0 ? (output.GetFirstTRef() + pos).GetJDay() : -999;
+		return pos >= 0 ? double((output.GetFirstTRef() + pos).GetJDay()) : -999.0;
 	}
